requirements.txt
python-dotenv==1.1.0
bleach==6.2.0
tenacity==9.1.2
PyTelegramBotAPI==4.27.0
pymongo==4.13.0
lxml==5.4.0
numpy>=1.21.0
tqdm==4.67.1
colorlog==6.9.0
python-telegram-bot==22.1
pandas>=1.5.0
langchain==0.3.25
langchain-core==0.3.65
langchain-elasticsearch==0.3.2
openpyxl>=3.0.0
psutil==7.0.0
aiohttp==3.12.12
scikit-learn>=1.0.0
matplotlib==3.10.3